Fixes browser width problem when moving splitter
With no maximum width set for the browser column, you could drag the
window splitter until the browser would size beyond the width of VS and
cover other windows.  The MaxWidth of the column is now set to 80% of
the total editor+browswer width.  I've also set the minimum width to be
150 pixels which keeps some odd behavior from happening when set to less
than that.