Add an option to set the HTML file extension (#168)
This option allows users to modify the file extension used in the generation of HTML files. The default remains .html, but this option allows the output to have a .cshtml extension. This allows documentation to be written in Markdown and a .cshtml file to be generated (combined with the HTML Template) to automatically produce web help pages for an MVC web application using the Razor template engine. It means there is no double handling or special code needed to integrate documentation into the web application - just write the Markdown and the a HTML Help file is generated and can be called by the appropriate Controller. 

The Razor engine takes care of the styling, layout etc. of the Help text.